added feature 'search patients via email-id and treat' for doctor

+++ hc/views.py
@@ -1,7 +1,8 @@
 from accounts.models import Doctor, Patient
-from .forms import takeAppointmentForm, treatPatientForm
+from .forms import takeAppointmentForm, treatPatientForm, SearchPatientForm
 from django.shortcuts import get_object_or_404, render, redirect
 from django.views.generic import CreateView
+from django.views.generic.base import View
 # from .forms import treatPatient
 import datetime as dt
 from.models import Appointment
@@ -27,8 +28,8 @@ def takeAppointmentView(request):
 class treatPatientView(CreateView):
     form_class = treatPatientForm
     template_name = 'hc/treat_patient.html'
-    # success_url = '/hc/treatPatient/'
     success_url = '/'
+    super_context = {}
 
     def get_context_data(self, *args, **kwargs):
         context = super(treatPatientView, self).get_context_data(*args, **kwargs)
@@ -36,18 +37,48 @@ class treatPatientView(CreateView):
             if(appointment.date < dt.date.today()):
                 appointment.delete()
             elif(appointment.date == dt.date.today()
-                 and appointment.time < (dt.datetime.now()-dt.timedelta(minutes=30)).time()):
+                 and appointment.time < (dt.datetime.now()-dt.timedelta(minutes=30)).time()
+                 and (dt.datetime.now()-dt.timedelta(minutes=30)).date() >= appointment.date):
                 appointment.delete()
             else:
                 break
-        appointment = Appointment.objects.filter().order_by('time')[0]
+        email = self.request.POST.get('email', False)
+        if email:
+            context['email'] = email
+            self.super_context['email'] = email
+            appointment = get_object_or_404(Appointment, patient__user__username=email)
+        else:
+            appointment = Appointment.objects.filter().order_by('time')[0]
         context['appointment'] = appointment
+        self.super_context['appointment'] = appointment
         return context
 
-    def form_valid(self, form):
-        appointment = Appointment.objects.filter().order_by('time')[0]
+    def form_valid(self, form, **kwargs):
+        context = self.super_context
+        if context['email']:
+            email = context['email']
+            appointment = get_object_or_404(Appointment, patient__user__username=email)
+        else:
+            appointment = Appointment.objects.filter().order_by('time')[0]
         prescription = form.save()
         prescription.doctor = appointment.doctor
         appointment.patient.prescriptions.add(prescription)
         appointment.delete()
         return super(treatPatientView, self).form_valid(form)
+
+
+class SearchPatientView(View):
+    form_class = SearchPatientForm()
+    template_name = 'hc/treat_patient.html'
+    success_url = 'treatPatient/'
+
+    def post(self, request, *args, **kwargs):
+        form = SearchPatientForm(request.POST)
+        if form.is_valid():
+            email = form['email'].value()
+            appointment = get_object_or_404(Appointment, patient__user__username=email)
+            prescription = form.save()
+            prescription.doctor = appointment.doctor
+            appointment.patient.prescriptions.add(prescription)
+            appointment.delete()
+            return super(SearchPatientView, self).form_valid(form)
